32bit floats again.

+++ src/libcolorscreen/include/color.h
@@ -4,7 +4,8 @@
 #include "matrix.h"
 
 //typedef double luminosity_t;
-typedef double luminosity_t;
+//typedef double luminosity_t;
+typedef float luminosity_t;
 /* mem_luminosity_t is used for very large temporary data.  */
 
 #ifdef __x86_64__
